remove editor padding on top and bottom
The issue of the tabs not reaching the top of the screen was due to padding in panes.less (it was padding of the entire editor pane rather than the tabs.

However, this raised an issue about padding on the right. When using minimap, it's no problem at all. Minimap is very happy sitting right up against the right edge of the screen when maximized. On the other hand, without minimap, the scrollbar is right up against the edge. This could become an issue for users who use their scrollbar (which shouldn't be many). If so, we can fix it then.

Fixes #14